+++ spec/controller_spec_helper.rb
@@ -16,24 +16,42 @@ module ControllerSpecHelper
 =end  
   ################
   ## sort
+=begin
   shared_context :sort_column do |key, direction = :asc, datatable:|
     it {
+      column = datatable.columns[key]
+
       get :index, params: { format: :json, sort_column: key, sort_direction: direction }
       got = JSON.parse(response.body).map_value(key.to_s).map(&:to_s)
-      expected = got.sort
-      expected.reverse! if direction.to_sym == :desc
-
-      column = datatable.columns[key]
 
       expected = datatable.data.order("#{column.source} #{direction}").map { |d| d.send(key).to_s }
       expect(got).to eq(expected)
     }
   end
-
+=end
   shared_context :sort do |datatable|
     datatable.columns.select(&:orderable).map(&:name).map(&:to_sym).each do |key|
-      include_context :sort_column, key, :asc, datatable: datatable
-      include_context :sort_column, key, :desc, datatable: datatable
+      it {
+        column = datatable.columns[key]
+
+        ## asc
+        direction = :asc
+        get :index, params: { format: :json, sort_column: key, sort_direction: direction }
+        got = JSON.parse(response.body).map_value(key.to_s).map(&:to_s)
+
+        expected = datatable.data.order("#{column.source} #{direction}").map { |d| d.send(key).to_s }
+        expect(got).to eq(expected)
+
+        ## desc
+        direction = :desc
+        get :index, params: { format: :json, sort_column: key, sort_direction: direction }
+        got = JSON.parse(response.body).map_value(key.to_s).map(&:to_s)
+
+        expected = datatable.data.order("#{column.source} #{direction}").map { |d| d.send(key).to_s }
+        expect(got).to eq(expected)
+      }
+      #include_context :sort_column, key, :asc, datatable: datatable
+      #include_context :sort_column, key, :desc, datatable: datatable
     end
   end