+++ spec/features/elements_feature_spec.rb
@@ -14,9 +14,12 @@ feature ElementsController, type: :feature, feature: true do
       it_behaves_like :contains, true, true
     end
     context 'filter' do
-      include_context :filter, ElementsDatatable::Filters.new,
-                      excludings: [:season_operator, :name_operator, :goe_operator]
+      include_context :filter, ElementsDatatable::Filters.new
       include_context :filter_season
+
+      conext 'element_name' do
+        #it_behaves_like :filter
+      end
 
       context 'element_name' do
         subject {