+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - sorting_sql: super
 - spec: update details
 - ajax helper: find by xpath
+- filters: flatten
 
 ### new feature